[DMenu] Huge list speedups (#1621)
* Don´t refilter on each key-press.

* Make sure refiltering is not completely starved.

* Force refilter before accepting entry

* Go into 'timeout' refilter mode only after certain # elements.

* [DMenu] threading with getdelim to speed up reading.

* [View] Remove debug output.

* [dmenu] Fix pre-read.

* [DMenu] Make sure that async mode cannot block

* Remove mutex for IPC via pipes.

* [Dmenu] Small cleanup.

* [Scrollbar] Fix scrollbar overflowing on big lists.

* Fix stack overflow by creating to many filter jobs.

* [Doc] Add new option to manpage

+++ source/widgets/scrollbar.c
@@ -161,9 +161,9 @@ void scrollbar_set_handle_length(scrollbar *sb, unsigned int pos_length) {
  */
 static void scrollbar_draw(widget *wid, cairo_t *draw) {
   scrollbar *sb = (scrollbar *)wid;
-  unsigned int wh = widget_padding_get_remaining_height(wid);
+  double wh = widget_padding_get_remaining_height(wid);
   // Calculate position and size.
-  unsigned int r = (sb->length * wh) / ((double)(sb->length + sb->pos_length));
+  double r = (sb->length * wh) / ((double)(sb->length + sb->pos_length));
   unsigned int handle = wid->h - r;
   double sec = ((r) / (double)(sb->length - 1));
   unsigned int height = handle;
